Only using the uppercase for LIKE ? queries. PK searches where broken.
- Bracket misplaced, which resulted on None vars also being attempted to upper.
- Added check to only upper SELECT statements. Up until now also INSERTS and UPDATES were upped. Resulting in all uppercased values inserted.

The ibm_db_django implementation creates tables in full uppercase. But when using the drivers in combination with an existing db2 database model, this could not be the case.
This patch, includes support for mixed lower/uppercase db tables.

The setting: CASE_INSENSITIVE_SEARCH, if not set, all searches against db2 need to be performed uppercased.